alterações nos métodos da controller e service

+++ api-codando/src/main/java/school/sptech/apicodando/service/educadorService/EducadorService.java
@@ -33,22 +33,19 @@ public class EducadorService {
     @Autowired
     private PasswordEncoder passwordEncoder;
 
-    private final GerenciadorTokenJwt gerenciadorTokenJwt;
+    @Autowired
+    private GerenciadorTokenJwt gerenciadorTokenJwt;
 
     @Autowired
     private AuthenticationManager authenticationManager;
 
-    @Autowired
-    public EducadorService(GerenciadorTokenJwt gerenciadorTokenJwt) {
-        this.gerenciadorTokenJwt = gerenciadorTokenJwt;
-    }
 
     public void criar(EducadorCadastroDTO educadorCadastroDTO) {
         final Educador novoEducador = EducadorMapper.toEntity(educadorCadastroDTO);
         String senhaCriptografada = passwordEncoder.encode(novoEducador.getSenha());
         novoEducador.setSenha(senhaCriptografada);
 
-        if (existePorEmail(educadorCadastroDTO.getEmail())){
+        if (existePorEmail(educadorCadastroDTO.getEmail())) {
             throw new ResponseStatusException(HttpStatus.CONFLICT, "Educador já cadastrado.");
         }
 
@@ -56,12 +53,21 @@ public class EducadorService {
     }
 
     public void excluir(int id) {
+
+        if (!existePorId(id)){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
+
         educadorRepository.deleteById(id);
     }
 
-    public void atualizar(Educador educadorAtualizado, int id){
-        educadorAtualizado.setIdEducador(id);
-        educadorRepository.save(educadorAtualizado);
+    public void atualizar(EducadorCadastroDTO educadorAtualizado, int id) {
+        if (!educadorRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Educador não encontrado.");
+        }
+        Educador educador = EducadorMapper.toEntity(educadorAtualizado);
+        educador.setIdEducador(id);
+        educadorRepository.save(educador);
     }
 
     public EducadorTokenDto autenticar(EducadorLoginDTO usuarioLoginDto) {
@@ -84,16 +90,18 @@ public class EducadorService {
         return EducadorMapper.of(usuarioAutenticado, token);
     }
 
-    public Optional<Educador> listarUmPorId(int id){
+    public Optional<Educador> listarUmPorId(int id) {
+        if (!educadorRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Educador não encontrado.");
+        }
         return educadorRepository.findById(id);
     }
 
-    public List<Educador> listarTodos(){
+    public List<Educador> listarTodos() {
+        if (educadorRepository.findAll().isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NO_CONTENT, "Nenhum educador cadastrado.");
+        }
         return educadorRepository.findAll();
-    }
-
-    public boolean existePorId(int id) {
-        return educadorRepository.existsById(id);
     }
 
     public Boolean existePorEmail(String email) {
@@ -101,6 +109,10 @@ public class EducadorService {
             return false;
         }
         return true;
+    }
+
+    public boolean existePorId(int id) {
+        return educadorRepository.findByIdEducador(id).isPresent();
     }
